created story context and migration scripts.

+++ src/GreenField.API/Modules/GreenField.Books/Data/DomainModels/Book.cs
@@ -14,7 +14,7 @@ namespace GreenField.Books.Data.DomainModels
             protected set { _bookEntityPictures = value; }
         }
         public string Name { get; set; }
-        public int AutherId { get; set; }
+        public long AutherId { get; set; }
         public virtual Author Author { get; set; }
     }
 }
